18w50d
1.8.0

+++ src/main/java/cn/nukkit/event/player/PlayerSettingsRespondedEvent.java
@@ -14,18 +14,18 @@ public class PlayerSettingsRespondedEvent extends PlayerEvent implements Cancell
         return handlers;
     }
 
-    protected int formID;
+    protected long formID;
     protected FormWindow window;
 
     protected boolean closed = false;
 
-    public PlayerSettingsRespondedEvent(Player player, int formID, FormWindow window) {
+    public PlayerSettingsRespondedEvent(Player player, long formID, FormWindow window) {
         this.player = player;
         this.formID = formID;
         this.window = window;
     }
 
-    public int getFormID() {
+    public long getFormID() {
         return this.formID;
     }
